+++ apps/api/internal/handlers/healthCheck.go
@@ -1,4 +1,4 @@
-package healthCheck
+package handlers
 
 import (
 	"encoding/json"
