+++ apps/api/internal/handlers/tenants.go
@@ -1,4 +1,4 @@
-package tenants
+package handlers
 
 import (
 	"encoding/base64"
@@ -6,6 +6,7 @@ import (
 
 	"github.com/QodeSrl/gardbase/apps/api/internal/services"
 	"github.com/QodeSrl/gardbase/apps/api/internal/storage"
+	"github.com/QodeSrl/gardbase/pkg/api/tenants"
 	"github.com/gin-gonic/gin"
 	"github.com/google/uuid"
 )
@@ -18,7 +19,7 @@ type TenantHandler struct {
 
 func (t *TenantHandler) HandleCreateTenant(c *gin.Context) {
 	tenantID := uuid.NewString()
-	var req CreateTenantRequest
+	var req tenants.CreateTenantRequest
 	if err := c.BindJSON(&req); err != nil {
 		c.JSON(400, gin.H{"error": fmt.Sprintf("Invalid create tenant request: %v", err)})
 		return
@@ -86,7 +87,7 @@ func (t *TenantHandler) HandleCreateTenant(c *gin.Context) {
 		return
 	}
 
-	c.JSON(200, CreateTenantResponse{
+	c.JSON(200, tenants.CreateTenantResponse{
 		TenantID: tenantID,
 		// TODO: later on, implement advanced self-managed keys
 		// EncryptedMasterKey:  res.Data.MasterKey,
